use_cases/TradingAgents-main-fa4d01c/stockbench/requirements.txt
pandas>=2.1.0,<2.3
numpy>=1.26,<2.1
pyarrow>=12.0.0
httpx[socks]>=0.24.0,<0.28.0
pydantic>=2.3.0,<3.0.0
typer>=0.9.0,<0.13.0
PyYAML>=6.0,<7.0
loguru>=0.7.0,<1.0.0
pytest>=7.0,<9.0
python-dotenv>=1.0.0,<2.0.0
finnhub-python>=2.4.0
openai>=1.64.0

# JSON repair tools
json-repair>=0.50.0
demjson3>=3.0.6

# Stock price analysis dependencies
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
requests>=2.31.0,<3.0.0

# Optional: Local inference (OpenAI compatible)
# - GPU version (requires CUDA drivers and VRAM):
# vllm>=0.10.0
# - CPU version (convenient for quick testing, slower speed):
# vllm-cpu>=0.5.5

